Added the corpus creator
This file takes two arguments when called: The root directory of where all Gutenberg Spanish text files are and the root directory of where all OCR-created newspaper text files are. It will create two corpora by scanning all text files and using all unique tokens, then compares them and creates a result text file. Within the result text file we have all words that were in the newspaper list and not in the Gutenberg list. This choice was because there are far more errors in the newspaper list which we can then look more closely at and correct.